// superset-frontend/plugins/plugin-chart-echarts/src/DodoExtensions/types.ts
// DODO was here
// DODO created 45525377
export enum LabelPositionDodo {
  Left = 'left',
  Right = 'right',
  Top = 'top',
  Bottom = 'bottom',
  Inside = 'inside',
  InsideTop = 'insideTop',
  InsideLeft = 'insideLeft',
  InsideRight = 'insideRight',
  InsideBottom = 'insideBottom',
  InsideTopLeft = 'insideTopLeft',
  InsideTopRight = 'insideTopRight',
  InsideBottomLeft = 'insideBottomLeft',
  InsideBottomRight = 'insideBottomRight',
}